requirements.txt
azure-functions
fastapi
playwright
beautifulsoup4